# Issue #3715 - Improve Log condensePackageString() performance