#3715
+ Adding alternate implementation
+ Adding more unit tests
+ Fixing original implementation
+ Removing Logger from StringUtil
+ Adding jmh test to compare implementations

Signed-off-by: Joakim Erdfelt <joakim.erdfelt@gmail.com>