@joakime 2019-05-31
JMH results of old implementation vs new implementation.

```
Benchmark                                                                                                                                                (fqClassName)   Mode  Cnt         Score         Error   Units
LogCondensePackageStringBenchmark.testFastCondense                                                                                                       com.acme.Dump  thrpt   20  16743321.094 ± 1032183.613   ops/s
LogCondensePackageStringBenchmark.testFastCondense:·gc.alloc.rate                                                                                        com.acme.Dump  thrpt   20      1509.541 ±      93.077  MB/sec
LogCondensePackageStringBenchmark.testFastCondense:·gc.alloc.rate.norm                                                                                   com.acme.Dump  thrpt   20       104.000 ±       0.001    B/op
LogCondensePackageStringBenchmark.testFastCondense:·gc.churn.PS_Eden_Space                                                                               com.acme.Dump  thrpt   20      1539.692 ±     253.136  MB/sec
LogCondensePackageStringBenchmark.testFastCondense:·gc.churn.PS_Eden_Space.norm                                                                          com.acme.Dump  thrpt   20       106.056 ±      17.183    B/op
LogCondensePackageStringBenchmark.testFastCondense:·gc.churn.PS_Survivor_Space                                                                           com.acme.Dump  thrpt   20         0.019 ±       0.012  MB/sec
LogCondensePackageStringBenchmark.testFastCondense:·gc.churn.PS_Survivor_Space.norm                                                                      com.acme.Dump  thrpt   20         0.001 ±       0.001    B/op
LogCondensePackageStringBenchmark.testFastCondense:·gc.count                                                                                             com.acme.Dump  thrpt   20       120.000                counts
LogCondensePackageStringBenchmark.testFastCondense:·gc.time                                                                                              com.acme.Dump  thrpt   20       509.000                    ms
LogCondensePackageStringBenchmark.testFastCondense                                   org.eclipse.jetty.websocket.common.extensions.compress.DeflateFrameExtension$Pool  thrpt   20   3804484.589 ±   81345.956   ops/s
LogCondensePackageStringBenchmark.testFastCondense:·gc.alloc.rate                    org.eclipse.jetty.websocket.common.extensions.compress.DeflateFrameExtension$Pool  thrpt   20       817.933 ±      17.492  MB/sec
LogCondensePackageStringBenchmark.testFastCondense:·gc.alloc.rate.norm               org.eclipse.jetty.websocket.common.extensions.compress.DeflateFrameExtension$Pool  thrpt   20       248.000 ±       0.001    B/op
LogCondensePackageStringBenchmark.testFastCondense:·gc.churn.PS_Eden_Space           org.eclipse.jetty.websocket.common.extensions.compress.DeflateFrameExtension$Pool  thrpt   20       833.717 ±      66.612  MB/sec
LogCondensePackageStringBenchmark.testFastCondense:·gc.churn.PS_Eden_Space.norm      org.eclipse.jetty.websocket.common.extensions.compress.DeflateFrameExtension$Pool  thrpt   20       252.663 ±      18.146    B/op
LogCondensePackageStringBenchmark.testFastCondense:·gc.churn.PS_Survivor_Space       org.eclipse.jetty.websocket.common.extensions.compress.DeflateFrameExtension$Pool  thrpt   20         0.030 ±       0.014  MB/sec
LogCondensePackageStringBenchmark.testFastCondense:·gc.churn.PS_Survivor_Space.norm  org.eclipse.jetty.websocket.common.extensions.compress.DeflateFrameExtension$Pool  thrpt   20         0.009 ±       0.004    B/op
LogCondensePackageStringBenchmark.testFastCondense:·gc.count                         org.eclipse.jetty.websocket.common.extensions.compress.DeflateFrameExtension$Pool  thrpt   20       216.000                counts
LogCondensePackageStringBenchmark.testFastCondense:·gc.time                          org.eclipse.jetty.websocket.common.extensions.compress.DeflateFrameExtension$Pool  thrpt   20       655.000                    ms
LogCondensePackageStringBenchmark.testOldCondense                                                                                                        com.acme.Dump  thrpt   20   2249153.335 ±   65390.871   ops/s
LogCondensePackageStringBenchmark.testOldCondense:·gc.alloc.rate                                                                                         com.acme.Dump  thrpt   20      2667.292 ±      77.554  MB/sec
LogCondensePackageStringBenchmark.testOldCondense:·gc.alloc.rate.norm                                                                                    com.acme.Dump  thrpt   20      1368.000 ±       0.001    B/op
LogCondensePackageStringBenchmark.testOldCondense:·gc.churn.PS_Eden_Space                                                                                com.acme.Dump  thrpt   20      2726.941 ±     277.190  MB/sec
LogCondensePackageStringBenchmark.testOldCondense:·gc.churn.PS_Eden_Space.norm                                                                           com.acme.Dump  thrpt   20      1400.147 ±     147.208    B/op
LogCondensePackageStringBenchmark.testOldCondense:·gc.churn.PS_Survivor_Space                                                                            com.acme.Dump  thrpt   20         0.023 ±       0.012  MB/sec
LogCondensePackageStringBenchmark.testOldCondense:·gc.churn.PS_Survivor_Space.norm                                                                       com.acme.Dump  thrpt   20         0.012 ±       0.006    B/op
LogCondensePackageStringBenchmark.testOldCondense:·gc.count                                                                                              com.acme.Dump  thrpt   20       113.000                counts
LogCondensePackageStringBenchmark.testOldCondense:·gc.time                                                                                               com.acme.Dump  thrpt   20       403.000                    ms
LogCondensePackageStringBenchmark.testOldCondense                                    org.eclipse.jetty.websocket.common.extensions.compress.DeflateFrameExtension$Pool  thrpt   20    798842.275 ±   15421.297   ops/s
LogCondensePackageStringBenchmark.testOldCondense:·gc.alloc.rate                     org.eclipse.jetty.websocket.common.extensions.compress.DeflateFrameExtension$Pool  thrpt   20      1266.483 ±      29.320  MB/sec
LogCondensePackageStringBenchmark.testOldCondense:·gc.alloc.rate.norm                org.eclipse.jetty.websocket.common.extensions.compress.DeflateFrameExtension$Pool  thrpt   20      1828.800 ±      22.807    B/op
LogCondensePackageStringBenchmark.testOldCondense:·gc.churn.PS_Eden_Space            org.eclipse.jetty.websocket.common.extensions.compress.DeflateFrameExtension$Pool  thrpt   20      1296.040 ±     190.148  MB/sec
LogCondensePackageStringBenchmark.testOldCondense:·gc.churn.PS_Eden_Space.norm       org.eclipse.jetty.websocket.common.extensions.compress.DeflateFrameExtension$Pool  thrpt   20      1870.309 ±     262.034    B/op
LogCondensePackageStringBenchmark.testOldCondense:·gc.churn.PS_Survivor_Space        org.eclipse.jetty.websocket.common.extensions.compress.DeflateFrameExtension$Pool  thrpt   20         0.022 ±       0.014  MB/sec
LogCondensePackageStringBenchmark.testOldCondense:·gc.churn.PS_Survivor_Space.norm   org.eclipse.jetty.websocket.common.extensions.compress.DeflateFrameExtension$Pool  thrpt   20         0.033 ±       0.021    B/op
LogCondensePackageStringBenchmark.testOldCondense:·gc.count                          org.eclipse.jetty.websocket.common.extensions.compress.DeflateFrameExtension$Pool  thrpt   20       127.000                counts
LogCondensePackageStringBenchmark.testOldCondense:·gc.time                           org.eclipse.jetty.websocket.common.extensions.compress.DeflateFrameExtension$Pool  thrpt   20       447.000                    ms
```

The highlights ...

| Name | Ops per Sec | Bytes allocated per Op |
| -------- | ------- | ------ |
| **Small Package** | | |
| Old Implementation | 2,249,153.335 ± 65390.871 | 1368 B |
| New Implementation | 16,743,321.094 ± 1032183.613 | 104 B |
| **Large Package** | | |
| Old Implementation | 798,842.275 ±   15421.297 | 1828 B |
| New Implementation | 3,804,484.589 ±   81345.956 | 248 B |